+++ src/main/java/de/cismet/belisEE/entity/Klassifizierung.java
@@ -28,14 +28,14 @@ public interface Klassifizierung extends Serializable {
      *
      * @return  DOCUMENT ME!
      */
-    Short getPk();
+    Integer getPk();
 
     /**
      * DOCUMENT ME!
      *
      * @param  pk  DOCUMENT ME!
      */
-    void setPk(final Short pk);
+    void setPk(final Integer pk);
 
     /**
      * DOCUMENT ME!
