+++ src/main/java/de/cismet/belisEE/entity/Leitung.java
@@ -38,14 +38,14 @@ public interface Leitung extends Serializable, DocumentContainer {
      *
      * @return  DOCUMENT ME!
      */
-    Long getId();
+    Integer getId();
 
     /**
      * DOCUMENT ME!
      *
      * @param  id  DOCUMENT ME!
      */
-    void setId(final Long id);
+    void setId(final Integer id);
 
     /**
      * DOCUMENT ME!
